[FE] refactor: redux관련 리팩토링
기존의 slice.js에서 tag 관련 항목을 뽑아서 tagSlice.js로 이동
